+++ CHANGE_LOG.md
@@ -1,3 +1,29 @@
+# [1.66.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.65.0...v1.66.0) (2022-11-18)
+
+
+### Bug Fixes
+
+* ensure SEO is pulled from listing pages ([9a6d6dc](https://github.com/oaknational/Oak-Web-Application/commit/9a6d6dcb73315b7a73e196495855dc44b53aee41))
+
+
+### Features
+
+* correct SEO props for webinar and blog listing pages ([82461b5](https://github.com/oaknational/Oak-Web-Application/commit/82461b5bcd5a59b9263b68a96b28cbec829c9d50))
+
+# [1.65.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.64.0...v1.65.0) (2022-11-17)
+
+
+### Features
+
+* category link to blog all blogs or webinars ([c42af21](https://github.com/oaknational/Oak-Web-Application/commit/c42af21ebb0f667be918f341d9d9f555927f2d81))
+
+# [1.64.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.63.0...v1.64.0) (2022-11-17)
+
+
+### Features
+
+* add webinars to homepage ([9d89a38](https://github.com/oaknational/Oak-Web-Application/commit/9d89a38f43de5e6c63619a204c90f0a6eee0d0e8))
+
 # [1.63.0](https://github.com/oaknational/Oak-Web-Application/compare/v1.62.1...v1.63.0) (2022-11-17)